requirements.txt
mysql-connector-python==9.2.0
pandas==2.2.2
numpy==1.26.4
tabulate==0.9.0
